+++ src/win.js
@@ -33,7 +33,9 @@ const binPowerShell = "powershell.exe";
  * @returns {string} The escaped argument.
  */
 function escapeArgCmd(arg, interpolation, quoted) {
-  let result = arg.replace(/\0/gu, "").replace(/\r?\n|\r/gu, " ");
+  let result = arg
+    .replace(/[\0\u0008\u001B\u009B]/gu, "")
+    .replace(/\r?\n|\r/gu, " ");
 
   if (interpolation) {
     result = result.replace(/\^/gu, "^^").replace(/(["&<>|])/gu, "^$1");
@@ -54,9 +56,10 @@ function escapeArgCmd(arg, interpolation, quoted) {
  */
 function escapeArgPowerShell(arg, interpolation, quoted) {
   let result = arg
-    .replace(/\0/gu, "")
+    .replace(/[\0\u0008\u001B\u009B]/gu, "")
     .replace(/`/gu, "``")
-    .replace(/\$/gu, "`$$");
+    .replace(/\$/gu, "`$$")
+    .replace(/\r(?!\n)/gu, "");
 
   if (interpolation) {
     result = result